% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcal.R
\name{ggcal}
\alias{ggcal}
\title{Given dates and a fill variable, generate a calendar plot}
\usage{
ggcal(dates, fills)
}
\arguments{
\item{dates}{vector of \code{Date} values}

\item{fills}{vector of values to map onto the fill aesthetic}
}
\description{
Given a collection of dates and values for each date, this
will generate a calendar for each month in the range of dates.
This will generate complete calendar months, so there is no need
to fill in any missing dates or anything.
}
\details{
Note that the ggplot2 object returned will not have any scale
defined for the fill, so that can be added to the returned
ggplot2 object.
}
\note{
If the dates span multiple years, the year will be appended
to the month name automatically, otherwise, it will not appear.
}
\examples{
{
mydate <- seq(as.Date("2017-02-01"), as.Date("2017-07-22"), by="1 day")
myfills <- rnorm(length(mydate))

print(ggcal(mydate, myfills))
}
}

